// src/features/congregation/publishers/types/enums.ts
export enum Gender {
  Male = "Male",
  Female = "Female",
}

export enum PublisherRole {
  Publisher = "Publisher",
  MinisterialServant = "Ministerial Servant",
  Elder = "Elder",
}

export enum PublisherType {
  Regular = "Regular Publisher",
  AuxiliaryPioneer = "Auxiliary Pioneer",
  RegularPioneer = "Regular Pioneer",
  SpecialPioneer = "Special Pioneer",
}

export enum PublisherStatus {
  Active = "Active",
  Inactive = "Inactive",
}